[finagle-core] Update OffloadFilter.Param API to encourage recommended construction
Problem:
The OffloadFilter.Param subclasses can be used for construction, but result in compiler errors when they are not specifically declared as type OffloadFilter.Param because the compiler is not able to infer the implicit value for their type.

Solution:
Reshape OffloadFilter.Param and remove the subclasses to avoid this issue.

JIRA Issues: CSL-11833

Differential Revision: https://phabricator.twitter.biz/D882210